backend/requirements.txt
django==5.1.6 
django-environ==0.12.0
django-cors-headers==4.7.0
djangorestframework==3.15.2
django-storages==1.14.4
psycopg==3.2.4
pillow==11.1.0